requirements.txt
# Core
streamlit>=1.35.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM clients
anthropic>=0.28.0
openai>=1.30.0

# Embeddings & retrieval
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
numpy>=1.26.0

# PDF processing
pdfplumber>=0.11.0
pypdf>=4.2.0

# Data
datasets>=2.19.0
huggingface-hub>=0.23.0

# Evaluation
scikit-learn>=1.4.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
PyYAML>=6.0.1
tiktoken>=0.7.0

# Testing
pytest>=8.2.0
